+++ src/Index.ts
@@ -15,4 +15,5 @@ export * from './TinodeHead.ts';
 export * from './TinodeLog.ts';
 export * from './TinodeMessage.ts';
 export * from './TinodeStorage.ts';
+export * from './TinodeTopics.ts';
 export * from './Tinode.ts';